+++ kttk/engines/abstract_engine.py
@@ -1,3 +1,5 @@
+import ktrack_api
+
 from kttk import logger
 from kttk.context import Context
 
@@ -30,6 +32,11 @@ class AbstractEngine(object):
         :return:
         """
         # extract contest from the workfile
+        # load full workfile
+        kt = ktrack_api.get_ktrack()
+
+        new_file = kt.find_one('workfile', new_file['id'])
+
         project = new_file['project']
         task = new_file['entity']
         step = task['step']
@@ -61,7 +68,7 @@ class AbstractEngine(object):
         Opens given workfile in DCC. Opened file will become current workfile. Will call change file
         :param file_to_open: workfile to open
         """
-        logger.info("Opening file {}".format(file_to_open['path']))
+        logger.info("Opening file {}".format(file_to_open))
         self.change_file(file_to_open)
 
 
@@ -95,6 +102,8 @@ class AbstractEngine(object):
         :return:
         """
         logger.info("Updating file for new context..")
+        logger.info("Serialize current context to file..")
+        self.serialize_context_to_file()
 
     def serialize_context_to_file(self):
         pass