+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- Maya Engine: serialize / deserialize context
+### Changed
+- Context: now only takes type and id on construction
 ## 0.2.0 - 2017-07-26
 ### Added
 - Added task_presets_manager: Can now get task presets for given entity type from config